test of userView display
